저장하기 추가

+++ src/main/kotlin/com/tobby/dailyapp/common/MessageResponse.kt
@@ -2,6 +2,6 @@ package com.tobby.dailyapp.common
 
 data class MessageResponse(
     val message: String,
-    val code: Int = 1,
+    val code: Long = 1,
     val reason: String? = null
 )
